requirements.txt
# Core scraping libraries
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Data handling
pandas>=2.1.0
numpy>=1.24.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0

# Logging and monitoring
structlog>=23.2.0
rich>=13.7.0

# Database support
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Avancerade funktioner - nya dependencies

# Redis för distribuerad scraping
aioredis>=2.0.0
redis>=5.0.0

# FastAPI för dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Machine learning för selektor-generering (valfritt)
scikit-learn>=1.3.0
scipy>=1.11.0

# Revolutionära AI-funktioner
openai>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# Avancerad HTTP-hantering
httpx>=0.25.0
aiofiles>=23.2.0

# Data serialisering
msgpack>=1.0.7
orjson>=3.9.0

# Monitoring och metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Avancerad parsing
parsel>=1.8.1
cssselect>=1.2.0

# Validering
pydantic>=2.5.0
cerberus>=1.3.0

# Caching
diskcache>=5.6.0

# Proxy-hantering
aiohttp-proxy>=0.1.2

# Webhook-hantering
webhooks>=0.0.1

# Plugin-system
importlib-metadata>=6.8.0
pluggy>=1.3.0
